split sector builder FFI from non-sector builder FFI
