+++ filecoin-proofs/src/api/mod.rs
@@ -1,10 +1,12 @@
-pub mod caches;
-pub mod constants;
-pub mod ffi;
-pub mod file_cleanup;
+mod caches;
+mod constants;
+mod file_cleanup;
 mod internal;
-pub mod parameters;
 mod post_adapter;
 mod responses;
 mod sector_builder;
+
+pub mod ffi_sector_builder;
+pub mod ffi_stateless;
+pub mod parameters;
 pub mod singletons;
